projects/voicepicker/docs/security/fastapi_users_reference.py
"""
Production-Grade Authentication System
Uses industry-standard libraries for secure authentication with 2FA support
"""

from fastapi import FastAPI, Depends, HTTPException, status
from fastapi_users import FastAPIUsers, BaseUserManager, UUIDIDMixin
from fastapi_users.authentication import (
    AuthenticationBackend,
    BearerTransport,
    JWTAuthentication,
    CookieTransport,
    RedisStrategy
)
from fastapi_users.db import SQLAlchemyAdapter
from sqlalchemy.ext.asyncio import AsyncSession, create_async_engine
from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column
from sqlalchemy import String, Boolean, DateTime
from passlib.context import CryptContext
import pyotp
import qrcode
from io import BytesIO
import base64
from datetime import datetime
from typing import Optional
import uuid

# Password hashing with Argon2 (more secure than bcrypt)
pwd_context = CryptContext(
    schemes=["argon2", "bcrypt"],  # Argon2 preferred, bcrypt fallback
    deprecated="auto",
    argon2__memory_cost=65536,  # 64 MB
    argon2__time_cost=3,        # 3 iterations
    argon2__parallelism=1,      # 1 thread
)

class Base(DeclarativeBase):
    pass

class User(Base):
    """Production User Model with 2FA support"""
    __tablename__ = "admin_users"
    
    id: Mapped[uuid.UUID] = mapped_column(primary_key=True, default=uuid.uuid4)
    email: Mapped[str] = mapped_column(String(320), unique=True, index=True)
    username: Mapped[str] = mapped_column(String(50), unique=True, index=True)
    hashed_password: Mapped[str] = mapped_column(String(1024))
    first_name: Mapped[Optional[str]] = mapped_column(String(100))
    last_name: Mapped[Optional[str]] = mapped_column(String(100))
    role: Mapped[str] = mapped_column(String(20), default="warehouse_manager")
    
    # Standard FastAPI-Users fields
    is_active: Mapped[bool] = mapped_column(Boolean, default=True)
    is_superuser: Mapped[bool] = mapped_column(Boolean, default=False)
    is_verified: Mapped[bool] = mapped_column(Boolean, default=False)
    
    # 2FA fields
    totp_secret: Mapped[Optional[str]] = mapped_column(String(32))
    is_2fa_enabled: Mapped[bool] = mapped_column(Boolean, default=False)
    backup_codes: Mapped[Optional[str]] = mapped_column(String(500))  # JSON array
    
    # Security fields
    last_login: Mapped[Optional[datetime]] = mapped_column(DateTime)
    failed_login_attempts: Mapped[int] = mapped_column(default=0)
    locked_until: Mapped[Optional[datetime]] = mapped_column(DateTime)
    password_changed_at: Mapped[Optional[datetime]] = mapped_column(DateTime)
    
    created_at: Mapped[datetime] = mapped_column(DateTime, default=datetime.utcnow)
    updated_at: Mapped[datetime] = mapped_column(DateTime, default=datetime.utcnow, onupdate=datetime.utcnow)

class UserManager(UUIDIDMixin, BaseUserManager[User, uuid.UUID]):
    """Enhanced User Manager with 2FA support"""
    reset_password_token_secret = "SECRET"  # Change in production
    verification_token_secret = "SECRET"    # Change in production

    async def on_after_register(self, user: User, request=None):
        """Called after user registration"""
        print(f"User {user.id} has registered.")

    async def on_after_login(self, user: User, request=None):
        """Called after successful login"""
        # Update last login time
        user.last_login = datetime.utcnow()
        user.failed_login_attempts = 0
        print(f"User {user.id} logged in.")

    async def on_after_forgot_password(self, user: User, token: str, request=None):
        """Called after password reset request"""
        print(f"User {user.id} has forgot their password. Reset token: {token}")

    def generate_totp_secret(self) -> str:
        """Generate TOTP secret for 2FA"""
        return pyotp.random_base32()

    def generate_qr_code(self, user: User, secret: str) -> str:
        """Generate QR code for 2FA setup"""
        totp_uri = pyotp.totp.TOTP(secret).provisioning_uri(
            name=user.email,
            issuer_name="VoicePicker Warehouse System"
        )
        
        qr = qrcode.QRCode(version=1, box_size=10, border=5)
        qr.add_data(totp_uri)
        qr.make(fit=True)
        
        img = qr.make_image(fill_color="black", back_color="white")
        buffer = BytesIO()
        img.save(buffer, format='PNG')
        buffer.seek(0)
        
        return base64.b64encode(buffer.getvalue()).decode()

    def verify_totp(self, secret: str, token: str) -> bool:
        """Verify TOTP token"""
        totp = pyotp.TOTP(secret)
        return totp.verify(token, valid_window=1)  # Allow 30s window

    def generate_backup_codes(self) -> list[str]:
        """Generate backup codes for 2FA"""
        import secrets
        return [secrets.token_hex(4).upper() for _ in range(10)]

# Authentication backends
bearer_transport = BearerTransport(tokenUrl="auth/jwt/login")
jwt_authentication = JWTAuthentication(
    secret="SECRET",  # Change in production
    lifetime_seconds=3600 * 24,  # 24 hours
    tokenUrl="auth/jwt/login",
)

# Cookie-based authentication (optional)
cookie_transport = CookieTransport(cookie_max_age=3600 * 24)
cookie_authentication = JWTAuthentication(
    secret="SECRET",  # Change in production
    lifetime_seconds=3600 * 24,
)

# Authentication backends
jwt_backend = AuthenticationBackend(
    name="jwt",
    transport=bearer_transport,
    get_strategy=lambda: jwt_authentication,
)

cookie_backend = AuthenticationBackend(
    name="cookie",
    transport=cookie_transport,
    get_strategy=lambda: cookie_authentication,
)

# FastAPI-Users setup
async def get_user_manager():
    # This would be properly implemented with your database
    pass

fastapi_users = FastAPIUsers[User, uuid.UUID](
    get_user_manager,
    [jwt_backend, cookie_backend],
)

# Dependency for current user
current_active_user = fastapi_users.current_user(active=True)
current_superuser = fastapi_users.current_user(active=True, superuser=True)

# 2FA Endpoints
from fastapi import APIRouter
from pydantic import BaseModel

router = APIRouter(prefix="/api/auth", tags=["Authentication"])

class Enable2FARequest(BaseModel):
    password: str

class Verify2FARequest(BaseModel):
    token: str

class Confirm2FARequest(BaseModel):
    token: str
    backup_codes: list[str]

@router.post("/2fa/setup")
async def setup_2fa(
    request: Enable2FARequest,
    user: User = Depends(current_active_user),
    user_manager: UserManager = Depends(get_user_manager)
):
    """Setup 2FA for user"""
    # Verify current password
    if not pwd_context.verify(request.password, user.hashed_password):
        raise HTTPException(status_code=400, detail="Invalid password")
    
    # Generate secret and QR code
    secret = user_manager.generate_totp_secret()
    qr_code = user_manager.generate_qr_code(user, secret)
    
    # Store secret temporarily (not enabled yet)
    user.totp_secret = secret
    
    return {
        "secret": secret,
        "qr_code": qr_code,
        "message": "Scan QR code with authenticator app, then verify with a token"
    }

@router.post("/2fa/verify")
async def verify_2fa_setup(
    request: Verify2FARequest,
    user: User = Depends(current_active_user),
    user_manager: UserManager = Depends(get_user_manager)
):
    """Verify and enable 2FA"""
    if not user.totp_secret:
        raise HTTPException(status_code=400, detail="2FA setup not initiated")
    
    # Verify token
    if not user_manager.verify_totp(user.totp_secret, request.token):
        raise HTTPException(status_code=400, detail="Invalid token")
    
    # Generate backup codes
    backup_codes = user_manager.generate_backup_codes()
    
    # Enable 2FA
    user.is_2fa_enabled = True
    user.backup_codes = ",".join(backup_codes)
    
    return {
        "message": "2FA enabled successfully",
        "backup_codes": backup_codes,
        "warning": "Save backup codes in a secure location"
    }

@router.post("/2fa/disable")
async def disable_2fa(
    request: Enable2FARequest,
    user: User = Depends(current_active_user)
):
    """Disable 2FA"""
    # Verify password
    if not pwd_context.verify(request.password, user.hashed_password):
        raise HTTPException(status_code=400, detail="Invalid password")
    
    # Disable 2FA
    user.is_2fa_enabled = False
    user.totp_secret = None
    user.backup_codes = None
    
    return {"message": "2FA disabled successfully"}

# Rate limiting middleware
from slowapi import Limiter, _rate_limit_exceeded_handler
from slowapi.util import get_remote_address
from slowapi.errors import RateLimitExceeded

limiter = Limiter(key_func=get_remote_address)

@router.post("/login")
@limiter.limit("5/minute")  # 5 login attempts per minute
async def enhanced_login(request, user_data: dict):
    """Enhanced login with rate limiting and 2FA"""
    # Login logic here
    pass

# Security headers middleware
from secure import Secure

secure_headers = Secure()

# Usage in your main app:
"""
from fastapi import FastAPI
app = FastAPI()

# Add security headers
@app.middleware("http")
async def add_security_headers(request, call_next):
    response = await call_next(request)
    secure_headers.framework.fastapi(response)
    return response

# Include router
app.include_router(router)
"""